ref: Move contacts tab from cabinet to separate common component
