+++ src/modules/common/Contacts/ContactsForm.styled.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { COLORS, FONTS, MEDIA } from '../../../theme';
+import { COLORS, FONTS, MEDIA } from '../../theme';
 
 export const Wrapper = styled.div`
   max-width: 1016px;
